refactor: extract WalletBalanceResult list converter
